requirements.txt
Django==3.2.3
djangorestframework
django-filter
requests
pytest
pytest-django
pytest-cov
django-environ
django-cors-headers
factory_boy
